# Update maxCommit for possible quorum intersection check performance improvements 