# Description

This PR will reduce maxCommit by 1 in the quorum intersection check algorithm.

## Correctness
Given a network with `N` nodes, the quorum intersection check algorithm essentially enumerates all the subsets of size `<= (N/2) + 1` and checks

* if the subset is a quorum, and
* if the complement contains a quorum.

It actually suffices to enumerate all the subsets of size `<= N / 2` because if there's a pair of disjoint quorums `A, B` with `|A| <= |B|`, then `|A| <= floor(N/2)`. (Otherwise, `|B| >= |A| >= floor(N/2) + 1` which implies `|A| + |B| >= 2 * (floor(N/2) + 1) = 2 * floor(N/2) + 2 > N`, which is a contradiction.)


## Possible performance improvements

This change will not make the program slower as we are reducing the search space.

Given a network of `N` nodes, our search space consists of

* `choose(N, 1)` subsets with 1 element,
* `choose(N, 2)` subsets with 2 elements,
* `choose(N, 3)` subsets with 3 elements,
* ...
* `choose(N, N/2 + 1)` subsets with `N/2 + 1` elements.

This change will get rid of the necessity of checking `choose(N, N/2 + 1)`. This could potentially be huge. For instance, when `N = 30`, the search space contains 759,852,346 sets, and this change would remove 145,422,675 sets, which is roughly 19%.

Furthermore, I did a quick verification in https://github.com/stellar/stellar-core/commit/94bbeebee2aca906bbeaeff58f00c16054966030, which shows that in one of the existing tests, we explore 40,310 sets, but this change would remove 17,722(=43.9643%) of them. (This number doesn't exactly match with the number above for any `N` because we have other early exits.)

# Checklist
- [x] Reviewed the [contributing](https://github.com/stellar/stellar-core/blob/master/CONTRIBUTING.md#submitting-changes) document
- [x] Rebased on top of master (no merge commits)
- [x] Ran `clang-format` v5.0.0 (via `make format` or the Visual Studio extension)
- [x] Compiles
- [x] Ran all tests
- [x] If change impacts performance, include supporting evidence per the [performance document](https://github.com/stellar/stellar-core/blob/master/performance-eval/performance-eval.md)