@hidenori-shinohara 2020-10-14
I did a quick performance test using `quorum intersection scaling test` in `src/herder/test/QuorumIntersectionTests.cpp` by running `time (src/stellar-core test 'quorum intersection scaling test')`.

I ran it 5 times on my local machine and

* Master branch => Average = 3.20 seconds, Median = 3.02 seconds.
* With this change =>  Average = 1.76 seconds, Median = 1.78 seconds.

which means that this change reduced the run time by around 40-45% in this particular test case.